refactor: one download path for community image endpoints (#13518)
* refactor: use the shared byte reader for community image downloads

readImageBuffer duplicated readResponseBytes, the helper every other media
download in the Worker uses, and MAX_IMAGE_BYTES duplicated MAX_IMAGE_SIZE.
Same 20 MB cap, same error, same 502.


* refactor: one download path for community image endpoints

The gen Worker and the enter registration probe each carried their own
copy of "pull the first image out of an OpenAI images response" — same
base64 branch, same URL branch, same 20 MB cap, drifted timeouts (120s
vs 90s) and two different byte readers.

- Hoists firstCommunityImageBytes() into shared/community-endpoints.ts.
  Callers pass an error factory so gen keeps throwing HttpError(502) and
  enter keeps throwing plain Error for the registration dialog.
- Moves readResponseBytes() to shared/ so the probe gets the streaming
  content-length + running-total cap instead of buffering arrayBuffer().
- Replaces both REQUEST_TIMEOUT_MS constants with one
  COMMUNITY_ENDPOINT_TIMEOUT_MS = 300s, matching the text providers
  (Portkey and Azure both use 290s).


* refactor: move HttpError to shared so both callers throw one class

The error-factory argument existed only because HttpError lived under
gen/src/image/, out of reach of enter. It is a plain Error subclass with
no gen-specific imports, so it moves to shared/ and the shared download
helper throws it directly.

enter loses nothing: throwEndpointTestError already flattens any Error to
HTTPException(400, {message}), so the 502 status is simply ignored there
while gen's funnel keeps rendering it.

38 of the 41 changed files are one-line import rewrites.


* style: sort imports after the HttpError move

Rewriting the specifier in place left @shared/http-error.ts in the old
sort position across the 38 mechanically-updated files.


---------
