+++ gen.pollinations.ai/src/embeddings/input.ts
@@ -1,6 +1,6 @@
 import { UpstreamError } from "@shared/error.ts";
+import { HttpError } from "@shared/http-error.ts";
 import { HTTPException } from "hono/http-exception";
-import { HttpError } from "@/image/httpError.ts";
 import { downloadImageAsBase64 } from "@/image/utils/imageDownload.ts";
 import { UserImageError } from "@/userImage.ts";
 import { MAX_EMBEDDING_BATCH_SIZE } from "./limits.ts";
