+++ gen.pollinations.ai/src/image/models/ideogramReplicateModel.ts
@@ -15,9 +15,9 @@
  * closest preset by log-space ratio distance.
  */
 
+import { HttpError } from "@shared/http-error.ts";
 import debug from "debug";
 import type { ImageGenerationResult } from "../createAndReturnImages.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { fetchUpstream } from "../utils/fetchUpstream.ts";
 import {
