+++ gen.pollinations.ai/src/image/models/minimaxH3Model.ts
@@ -1,6 +1,6 @@
+import { HttpError } from "@shared/http-error.ts";
 import type { VideoGenerationResult } from "../createAndReturnVideos.ts";
 import { getImageEnv } from "../env.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { sleep } from "../util.ts";
 import { fetchUpstream } from "../utils/fetchUpstream.ts";
