+++ gen.pollinations.ai/src/image/models/novaCanvasModel.ts
@@ -1,6 +1,6 @@
+import { HttpError } from "@shared/http-error.ts";
 import debug from "debug";
 import { getImageEnv } from "../env.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { base64ToBuffer, downloadUserImage } from "../utils/imageDownload.ts";
