+++ gen.pollinations.ai/src/image/models/novaReelModel.ts
@@ -1,6 +1,6 @@
+import { HttpError } from "@shared/http-error.ts";
 import debug from "debug";
 import { getImageEnv } from "../env.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { sleep } from "../util.ts";
 import { downloadUserImage } from "../utils/imageDownload.ts";
