+++ gen.pollinations.ai/src/image/models/seedanceV2VideoModel.ts
@@ -6,11 +6,11 @@
  * reference_videos, which would trigger Replicate's "video_in" price tier.
  */
 
+import { HttpError } from "@shared/http-error.ts";
 import { IMAGE_SERVICES } from "@shared/registry/image.ts";
 import type { ModelDefinition } from "@shared/registry/registry.ts";
 import debug from "debug";
 import type { VideoGenerationResult } from "../createAndReturnVideos.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { fetchUpstream } from "../utils/fetchUpstream.ts";
 import { toDataUri } from "../utils/imageDownload.ts";
