+++ gen.pollinations.ai/src/image/models/seedreamReplicateModel.ts
@@ -14,9 +14,9 @@
  * mode (only seedream 4.0 does).
  */
 
+import { HttpError } from "@shared/http-error.ts";
 import debug from "debug";
 import type { ImageGenerationResult } from "../createAndReturnImages.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { closestRatioLogSpace } from "../utils/aspectRatio.ts";
 import { fetchUpstream } from "../utils/fetchUpstream.ts";
