+++ gen.pollinations.ai/src/image/models/wanImageModel.ts
@@ -11,9 +11,9 @@
  * at the 2K tier (4K for pro text-to-image, where the upstream allows it).
  */
 
+import { HttpError } from "@shared/http-error.ts";
 import debug from "debug";
 import type { ImageGenerationResult } from "../createAndReturnImages.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { closestByRatio } from "../utils/aspectRatio.ts";
 import { fetchUpstream } from "../utils/fetchUpstream.ts";
