+++ gen.pollinations.ai/src/image/models/zImageFalModel.ts
@@ -1,6 +1,6 @@
+import { HttpError } from "@shared/http-error.ts";
 import type { ImageGenerationResult } from "../createAndReturnImages.ts";
 import { getImageEnv } from "../env.ts";
-import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import { fetchUpstream } from "../utils/fetchUpstream.ts";
