+++ gen.pollinations.ai/test/image/gptImageRouting.test.ts
@@ -1,10 +1,10 @@
+import type { HttpError } from "@shared/http-error.ts";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import {
     type AuthResult,
     callGPTImage,
 } from "../../src/image/createAndReturnImages.ts";
 import { syncImageEnv } from "../../src/image/env.ts";
-import type { HttpError } from "../../src/image/httpError.ts";
 import type { ImageParams } from "../../src/image/params.ts";
 
 const AZURE_KEY_ENV = {
