+++ shared/response-bytes.ts
